Run misspell self-hosted & from maintained repo
Run it with our own machine & take it from the repo that
is maintained.

However, as it's a self-hosted machine - the lib is already
there and it's also initialized. As are the node modules. These
contain LOTS of typos. Thus we must now specify which files/folders
to check (yes, it is a pain).